renamed to AudioVideoEvidenceUtil
